+++ server_2x/src/auth/interfaces/jwt-options.model.ts
@@ -1,4 +1,0 @@
-export interface JwtOptions {
-  expiresIn: number;
-  secret: string;
-}