+++ server_2x/src/crypto/crypto.module.ts
@@ -1,8 +1,0 @@
-import { Module } from "@nestjs/common";
-import { CryptoService } from "./crypto.service";
-
-@Module({
-  providers: [CryptoService],
-  exports: [CryptoService]
-})
-export class CryptoModule {}