+++ server_2x/src/dashboard/dto/sse-message.dto.ts
@@ -1,8 +1,0 @@
-import { ClientDashboardSettings } from "../../settings/models/client-dashboard/client-dashboard.settings";
-
-export interface DashboardSseMessageDto {
-  printerInformation: any;
-  currentOperations: any;
-  dashStatistic: any;
-  dashboardSettings: ClientDashboardSettings;
-}