+++ server_2x/src/printers/dto/create-multiple-printers.dtot.ts
@@ -1,3 +1,0 @@
-import { CreatePrinterDto } from "./create-printer.dto";
-
-export type CreateMultiplePrintersDtot = CreatePrinterDto[];