+++ server_2x/src/printers/dto/printers-sse-message.dto.ts
@@ -1,5 +1,0 @@
-export interface PrintersSseMessageDto {
-  printersInformation: any;
-  printerControlList: any;
-  currentTickerList: any;
-}