+++ server_2x/src/printers/utils/property.util.ts
@@ -1,3 +1,0 @@
-export function toDefinedKeyValue(prop, key) {
-  return typeof prop !== "undefined" && prop !== null ? { [key]: prop } : {};
-}