+++ server_2x/src/settings/models/client-dashboard/printers-states.model.ts
@@ -1,7 +1,0 @@
-export interface PrinterStatesModel {
-  printerState: boolean;
-  printerTemps: boolean;
-  printerUtilisation: boolean;
-  printerProgress: boolean;
-  currentStatus: boolean;
-}