+++ server_2x/src/settings/models/layout-settings/layout.model.ts
@@ -1,8 +1,0 @@
-import { LayoutEntryModel } from "./layout-entry.model";
-
-// Note: can be expanded to interface, but will break some code (which is fine)
-export type LayoutModel = LayoutEntryModel[];
-
-// export interface LayoutModel {
-//     layout: LayoutEntryModel[];
-// }