+++ server_2x/src/settings/models/server/server.model.ts
@@ -1,5 +1,0 @@
-export interface ServerModel {
-  registration: boolean;
-  port: number;
-  loginRequired: boolean;
-}