+++ server_2x/src/utils/property-reflection.util.ts
@@ -1,1 +1,0 @@
-export const nameof = <T>(name: keyof T): string => name as string;