+++ server_src/constants/server-message.constants.js
@@ -1,6 +1,6 @@
 module.exports = {
   SERVER_ISSUES: {
-    SYSTEM_NOT_SETUP: "System has not been setup.. loading setup screen...",
+    SYSTEM_NOT_SETUP: "Admin has not been setup.. loading setup screen...",
     DATABASE_AUTH_FAIL: "Database authentication failed...",
     DATABASE_CONN_FAIL: "Database connection failed..."
   }